requirements-dev.txt
apache-airflow==2.0.2
SQLAlchemy==1.3.23 # only needed locally because 1.4.0 break initilizing the DB
flake8==3.8.4
pytest==6.2.2
coverage==5.4
